requirements.txt
# ===== Web Framework =====
Django==5.1.2
django-ninja==1.2.4

# ===== Whisper STT =====
faster-whisper==1.2.1
ctranslate2==4.6.1
tokenizers==0.22.1
sentencepiece==0.2.1

# ===== PyTorch (CPU-only) =====
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
-f https://download.pytorch.org/whl/cpu

# ===== Audio Processing =====
librosa==0.11.0
soundfile==0.13.1
soxr==1.0.0
pydub==0.25.1
av==16.0.1

# ===== Utilities =====
numpy==2.3.1
requests==2.32.5
PyYAML==6.0.2
python-dateutil==2.9.0.post0
tiktoken==0.12.0
urllib3==2.5.0

